requirements.txt
streamlit
langchain
chromadb
openai
python-dotenv
PyPDF2
unstructured
